pyproject.toml
[project]
name = "stream-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.133.1",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
    "vision-agents[cartesia,deepgram,gemini,getstream,openrouter,smart-turn]>=0.3.8",
]
